refactor(agent): extract tool_execution + tool_definitions
Finish the agent.py decomposition started by #382 / #407 by relocating
the tool-invocation and tool-registry clusters into dedicated modules:

- New `tool_definitions.py` (166 lines): `collect_all_tool_defs`,
  `build_tool_list`, `refresh_dynamic_tools`, `invalidate_skill_cache`,
  and the `_skill_def_cache` they share.
- New `tool_execution.py` (364 lines): `execute_tool_calls`,
  `execute_single_tool`, `process_tool_media`, `resolve_widget`, and
  the `_media_placeholder_pattern` LRU helper.

Leading underscores dropped on the relocated public functions; module-
internal helpers stay private. `_archive`, `_check_cancelled`, and
`_conv_id` (22 lines) are duplicated into `tool_execution.py` rather
than extracted to a third shared module — they're trivial leaves the
agent loop calls into, and a comment in `tool_execution.py` documents
the intentional duplication.

`context_composer.py` now imports `collect_all_tool_defs` from
`.tool_definitions` instead of reaching into `agent.py` — preparing
the ground for #439. `tools/skill_tools.py` likewise updated to
import `invalidate_skill_cache` from the new module.

CLAUDE.md "Core" key-files list and `docs/architecture.md`
tool-execution snippet updated to reflect the new module layout.

agent.py: 1451 -> 1000 lines. The ~250 lines beyond the spec's "~750"
target are the wiki/attachment helpers (`_resolve_attachments`,
`_parse_wiki_references`, `_read_wiki_page`,
`_get_already_injected_pages`, `_WIKI_MENTION_RE`) explicitly deferred
to issue #439 per the spec's "What we're NOT doing" section.

Closes #438
