+++ tests/test_orphaned_tool_results.py
@@ -113,7 +113,7 @@ def _compose(ctx, config, user_message, history):
     config.agent.tool_context_budget_pct = 1.0
     config.compaction.max_tokens = 1000000
     with (
-        patch("decafclaw.agent._collect_all_tool_defs", return_value=[]),
+        patch("decafclaw.tool_definitions.collect_all_tool_defs", return_value=[]),
         patch("decafclaw.memory_context.retrieve_memory_context",
               new_callable=AsyncMock, return_value=[]),
     ):
